Refactored issues
